+++ tokio-trace-core/src/callsite.rs
@@ -7,7 +7,7 @@ use std::{
     sync::Mutex,
 };
 use {
-    dispatcher::{self, Dispatch},
+    dispatcher::{self, Dispatch, Registrar},
     subscriber::Interest,
     Metadata,
 };
@@ -24,23 +24,39 @@ struct Registry {
     dispatchers: Vec<dispatcher::Registrar>,
 }
 
-/// Trait implemented by callsites.
-pub trait Callsite: Sync {
-    /// Adds the [`Interest`] returned by [registering] the callsite with a
-    /// [dispatcher].
-    ///
-    /// If the interest is greater than or equal to the callsite's current
-    /// interest, this should change whether or not the callsite is enabled.
-    ///
-    /// [`Interest`]: ../subscriber/struct.Interest.html
-    /// [registering]: ../subscriber/trait.Subscriber.html#method.register_callsite
-    /// [dispatcher]: ../dispatcher/struct.Dispatch.html
-    fn add_interest(&self, interest: Interest);
+impl Registry {
+    fn rebuild_callsite_interest(&self, callsite: &'static Callsite) {
+        let meta = callsite.metadata();
 
-    /// Remove _all_ [`Interest`] from the callsite, disabling it.
+        let mut interest = Interest::never();
+
+        for registrar in &self.dispatchers {
+            if let Some(sub_interest) = registrar.try_register(meta) {
+                interest = interest.and(sub_interest);
+            }
+        }
+
+        callsite.set_interest(interest)
+    }
+
+    fn rebuild_interest(&mut self) {
+        self.dispatchers.retain(Registrar::is_alive);
+
+        self.callsites.iter().for_each(|&callsite| {
+            self.rebuild_callsite_interest(callsite);
+        });
+    }
+}
+
+/// Trait implemented by callsites.
+///
+/// These functions are only intended to be called by the [`Registry`] which
+/// correctly handles determining the common interest between all subscribers.
+pub trait Callsite: Sync {
+    /// Sets the [`Interest`] for this callsite.
     ///
     /// [`Interest`]: ../subscriber/struct.Interest.html
-    fn clear_interest(&self);
+    fn set_interest(&self, interest: Interest);
 
     /// Returns the [metadata] associated with the callsite.
     ///
@@ -67,34 +83,37 @@ pub struct Identifier(
     pub &'static Callsite,
 );
 
+/// Clear and reregister interest on every [`Callsite`]
+///
+/// This function is intended for runtime reconfiguration of filters on traces
+/// when the filter recalculation is much less frequent than trace events are.
+/// The alternative is to have the [`Subscriber`] that supports runtime
+/// reconfiguration of filters always return [`Interest::sometimes()`] so that
+/// [`enabled`] is evaluated for every event.
+///
+/// [`Callsite`]: ../callsite/trait.Callsite.html
+/// [`enabled`]: ../subscriber/trait.Subscriber.html#tymethod.enabled
+/// [`Interest::sometimes()`]: ../subscriber/struct.Interest.html#method.sometimes
+/// [`Subscriber`]: ../subscriber/trait.Subscriber.html
+pub fn rebuild_interest_cache() {
+    let mut registry = REGISTRY.lock().unwrap();
+    registry.rebuild_interest();
+}
+
 /// Register a new `Callsite` with the global registry.
 ///
 /// This should be called once per callsite after the callsite has been
 /// constructed.
 pub fn register(callsite: &'static Callsite) {
     let mut registry = REGISTRY.lock().unwrap();
-    let meta = callsite.metadata();
-    registry.dispatchers.retain(|registrar| {
-        match registrar.try_register(meta) {
-            Some(interest) => {
-                callsite.add_interest(interest);
-                true
-            }
-            // TODO: if the dispatcher has been dropped, should we invalidate
-            // any callsites that it previously enabled?
-            None => false,
-        }
-    });
+    registry.rebuild_callsite_interest(callsite);
     registry.callsites.push(callsite);
 }
 
 pub(crate) fn register_dispatch(dispatch: &Dispatch) {
     let mut registry = REGISTRY.lock().unwrap();
     registry.dispatchers.push(dispatch.registrar());
-    for callsite in &registry.callsites {
-        let interest = dispatch.register_callsite(callsite.metadata());
-        callsite.add_interest(interest);
-    }
+    registry.rebuild_interest();
 }
 
 // ===== impl Identifier =====
