+++ tokio-trace-core/src/subscriber.rs
@@ -19,7 +19,7 @@ use std::{
 /// - Registering new spans as they are created, and providing them with span
 ///   IDs. Implicitly, this means the subscriber may determine the strategy for
 ///   determining span equality.
-/// - Visiting the attachment of field values and follows-from annotations to
+/// - Recording the attachment of field values and follows-from annotations to
 ///   spans.
 /// - Filtering spans and events, and determining when those filters must be
 ///   invalidated.
@@ -50,9 +50,9 @@ pub trait Subscriber: 'static {
     /// indicate different interests, or to implement behaviour that should run
     /// once for every callsite.
     ///
-    /// This function is guaranteed to be called exactly once per callsite on
+    /// This function is guaranteed to be called at least once per callsite on
     /// every active subscriber. The subscriber may store the keys to fields it
-    /// cares in order to reduce the cost of accessing fields by name,
+    /// cares about in order to reduce the cost of accessing fields by name,
     /// preallocate storage for that callsite, or perform any other actions it
     /// wishes to perform once for each callsite.
     ///
@@ -79,6 +79,10 @@ pub trait Subscriber: 'static {
     /// set of metadata. Thus, the counter will not be incremented, and the span
     /// or event that correspands to the metadata will never be `enabled`.
     ///
+    /// `Subscriber`s that need to change their filters occasionally should call
+    /// [`rebuild_interest_cache`] to re-evaluate `register_callsite` for all
+    /// callsites.
+    ///
     /// Similarly, if a `Subscriber` has a filtering strategy that can be
     /// changed dynamically at runtime, it would need to re-evaluate that filter
     /// if the cached results have changed.
@@ -91,14 +95,19 @@ pub trait Subscriber: 'static {
     /// return `Interest::Never`, as a new subscriber may be added that _is_
     /// interested.
     ///
-    /// **Note**: If a subscriber returns `Interest::never` for a particular
-    /// callsite, it _may_ still see spans and events originating from that
-    /// callsite, if another subscriber expressed interest in it.
+    /// # Notes
+    /// This function may be called again when a new subscriber is created or
+    /// when the registry is invalidated.
+    ///
+    /// If a subscriber returns `Interest::never` for a particular callsite, it
+    /// _may_ still see spans and events originating from that callsite, if
+    /// another subscriber expressed interest in it.
     ///
     /// [filter]: #method.enabled
     /// [metadata]: ../metadata/struct.Metadata.html
     /// [`Interest`]: struct.Interest.html
     /// [`enabled`]: #method.enabled
+    /// [`rebuild_interest_cache`]: ../callsite/fn.rebuild_interest_cache.html
     fn register_callsite(&self, metadata: &Metadata) -> Interest {
         match self.enabled(metadata) {
             true => Interest::always(),
@@ -396,7 +405,7 @@ impl Interest {
     ///
     /// If all active subscribers are `sometimes` or `never` interested in a
     /// callsite, the currently active subscriber will be asked to filter that
-    /// callsite every time it creates a span. This will be the case until a
+    /// callsite every time it creates a span. This will be the case until a new
     /// subscriber expresses that it is `always` interested in the callsite.
     #[inline]
     pub fn sometimes() -> Self {
@@ -440,6 +449,29 @@ impl Interest {
         match self.0 {
             InterestKind::Always => true,
             _ => false,
+        }
+    }
+
+    /// Returns the common interest between these two Interests.
+    ///
+    /// The common interest is defined as the least restrictive, so if one
+    /// interest is `never` and the other is `always` the common interest is
+    /// `always`.
+    pub(crate) fn and(self, rhs: Interest) -> Self {
+        match rhs.0 {
+            // If the added interest is `never()`, don't change anything —
+            // either a different subscriber added a higher interest, which we
+            // want to preserve, or the interest is 0 anyway (as it's
+            // initialized to 0).
+            InterestKind::Never => self,
+            // If the interest is `sometimes()`, that overwrites a `never()`
+            // interest, but doesn't downgrade an `always()` interest.
+            InterestKind::Sometimes if self.0 == InterestKind::Never => rhs,
+            // If the interest is `always()`, we overwrite the current interest,
+            // as always() is the highest interest level and should take
+            // precedent.
+            InterestKind::Always => rhs,
+            _ => self,
         }
     }
 }
